wifi_battle_tower dirs added, files renamed and reorganised
