--- include/overlay096/struct_ov96_0223BF40_decl.h
@@ -1,6 +1,0 @@
-#ifndef POKEPLATINUM_STRUCT_OV96_0223BF40_DECL_H
-#define POKEPLATINUM_STRUCT_OV96_0223BF40_DECL_H
-
-typedef struct UnkStruct_ov96_0223BF40_t WifiBattleTowerAppState;
-
-#endif // POKEPLATINUM_STRUCT_OV96_0223BF40_DECL_H
