+++ platinum.us/sbins_shared.sha1
@@ -94,7 +94,7 @@ de8a847bff8c343d69b853a215e6ee775ef2ef96 *dummy6.sbin
 87c3ffad8da5d31174a542832d80bf1855c3225c *overlay92.sbin
 dabd69f841fce75c91c4236925757150990af289 *cutscenes.sbin
 3410ea804433b4a797932d1999fbb2429611ae56 *overlay95.sbin
-af7de8e345ff25de55130069f313b57e85d72a4f *overlay96.sbin
+af7de8e345ff25de55130069f313b57e85d72a4f *wifi_battle_tower.sbin
 138a69ea796d35f244daa36a7e9962ea2d0322e0 *main_menu.sbin
 7648d3a8ca287175a5e29aef623156f11c632d98 *overlay98.sbin
 e93f9eb3382856feb76c8b0ba400b7d2593db5c5 *overlay99.sbin
