+++ src/meson.build
@@ -120,7 +120,7 @@ pokeplatinum_c = files(
     'play_time.c',
     'chatot_cry.c',
     'game_records.c',
-    'wifi_battle_tower_save.c',
+    'wifi_battle_tower/save.c',
     'unk_0202D778.c',
     'special_encounter.c',
     'global_trade.c',
@@ -868,9 +868,9 @@ pokeplatinum_c = files(
     'overlay095/ov95_0224A390.c',
     'overlay095/ov95_0224ABC0.c',
     'overlay095/ov95_0224B3D8.c',
-    'overlay096/ov96_0223B140.c',
-    'overlay096/ov96_0223B6A0.c',
-    'overlay096/ov96_0223BCE0.c',
+    'wifi_battle_tower/http.c',
+    'wifi_battle_tower/application.c',
+    'wifi_battle_tower/screen.c',
     'main_menu/main_menu.c',
     'main_menu/ranger_link.c',
     'main_menu/ov97_0222D04C.c',
