+++ src/unk_0206B9D8.c
@@ -13,6 +13,7 @@
 #include "applications/party_menu/main.h"
 #include "applications/pokemon_summary_screen/main.h"
 #include "field/field_system.h"
+#include "wifi_battle_tower/save.h"
 
 #include "bag.h"
 #include "battle_frontier_save.h"
@@ -32,7 +33,6 @@
 #include "unk_02038FFC.h"
 #include "unk_0203D1B8.h"
 #include "unk_0204AEE8.h"
-#include "wifi_battle_tower_save.h"
 
 #include "constdata/const_020F410C.h"
