+++ src/wifi_battle_tower/http.c
@@ -1,4 +1,4 @@
-#include "overlay096/ov96_0223B140.h"
+#include "wifi_battle_tower/http.h"
 
 #include <dwc.h>
 #include <nitro.h>
@@ -7,10 +7,10 @@
 #include "struct_defs/wifi_battle_tower_data.h"
 
 #include "http/http.h"
-#include "overlay096/ov96_0223BCE0.h"
 #include "overlay096/struct_wifi_player_profile.h"
 #include "overlay096/struct_world_exchange_trainer.h"
 #include "overlay096/struct_world_exchange_trainer_error.h"
+#include "wifi_battle_tower/screen.h"
 
 #include "enums.h"
